+++ src/views/studio/StudioLayout.tsx
@@ -1,9 +1,8 @@
 import styled from '@emotion/styled'
 import { ErrorBoundary } from '@sentry/react'
-import React, { useEffect, useRef, useState } from 'react'
+import React, { useEffect, useState } from 'react'
 import { Route, Routes } from 'react-router'
 import { useLocation, useNavigate } from 'react-router-dom'
-import shallow from 'zustand/shallow'
 
 import { NoConnectionIndicator } from '@/components/NoConnectionIndicator'
 import { PrivateRoute } from '@/components/PrivateRoute'
@@ -18,12 +17,9 @@ import { ConnectionStatusManager, useConnectionStatusStore } from '@/providers/c
 import { useDialog } from '@/providers/dialogs'
 import { EditVideoSheetProvider, useVideoEditSheetRouting } from '@/providers/editVideoSheet'
 import { JoystreamProvider } from '@/providers/joystream'
-import { useSnackbar } from '@/providers/snackbars'
 import { TransactionManager } from '@/providers/transactionManager'
-import { UploadsManager, useUploadsStore } from '@/providers/uploadsManager'
-import { AssetUpload } from '@/providers/uploadsManager/types'
+import { UploadsManager } from '@/providers/uploadsManager'
 import { ActiveUserProvider, useUser } from '@/providers/user'
-import { SvgGlyphExternal } from '@/shared/icons'
 import { isAllowedBrowser } from '@/utils/browser'
 import {
   CreateEditChannelView,
@@ -35,33 +31,15 @@ import {
   SignInView,
 } from '@/views/studio'
 
-type UploadStatus = 'completed' | 'inProgress' | 'error' | 'reconnecting' | 'processing'
-type VideoAssets = AssetUpload & { uploadStatus?: UploadStatus }
-
 const ENTRY_POINT_ROUTE = absoluteRoutes.studio.index()
-const UPLOADED_SNACKBAR_TIMEOUT = 13000
 
 const StudioLayout = () => {
   const location = useLocation()
-  const navigate = useNavigate()
   const displayedLocation = useVideoEditSheetRouting()
   const internetConnectionStatus = useConnectionStatusStore((state) => state.internetConnectionStatus)
   const nodeConnectionStatus = useConnectionStatusStore((state) => state.nodeConnectionStatus)
-  const { displaySnackbar } = useSnackbar()
-  const videoAssetsRef = useRef<VideoAssets[]>([])
   const { activeAccountId, activeMemberId, activeChannelId, extensionConnected, memberships, userInitialized } =
     useUser()
-  const { assetsFiles, channelUploads, uploadStatuses } = useUploadsStore(
-    (state) => ({
-      assetsFiles: state.assetsFiles,
-      channelUploads: state.uploads.filter((asset) => asset.owner === activeChannelId),
-      uploadStatuses: state.uploadsStatus,
-    }),
-    shallow
-  )
-  const videoAssets = channelUploads
-    .filter((asset) => asset.type === 'video')
-    .map((asset) => ({ ...asset, uploadStatus: uploadStatuses[asset.contentId]?.lastStatus }))
 
   const [openUnsupportedBrowserDialog, closeUnsupportedBrowserDialog] = useDialog()
   const [enterLocation] = useState(location.pathname)
@@ -70,31 +48,6 @@ const StudioLayout = () => {
   const accountSet = !!activeAccountId && !!extensionConnected
   const memberSet = accountSet && !!activeMemberId && hasMembership
   const channelSet = memberSet && !!activeChannelId && hasMembership
-
-  // display snackbar when video upload is complete
-  useEffect(() => {
-    if (videoAssets.length) {
-      videoAssets.forEach((video) => {
-        const videoObject = videoAssetsRef.current.find(
-          (videoRef) => videoRef.uploadStatus !== 'completed' && videoRef.contentId === video.contentId
-        )
-        if (videoObject && video.uploadStatus === 'completed') {
-          const file = assetsFiles.find((asset) => asset.contentId === video.contentId)
-          displaySnackbar({
-            customId: video.contentId,
-            title: 'Video ready to be reviewed',
-            description: (file?.blob as File).name,
-            iconType: 'success',
-            timeout: UPLOADED_SNACKBAR_TIMEOUT,
-            actionText: 'See on Joystream',
-            actionIcon: <SvgGlyphExternal />,
-            onActionClick: () => window.open(`/video/${video.parentObject.id}`, '_blank'),
-          })
-        }
-      })
-      videoAssetsRef.current = videoAssets
-    }
-  }, [assetsFiles, displaySnackbar, navigate, videoAssets])
 
   useEffect(() => {
     if (!isAllowedBrowser()) {
